LPD-71138 Rename propsTransformer
