+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/publications/js/components/OngoingPublicationsFDSPropsTransformer.js
