+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/publications/view_publications.jsp
@@ -23,7 +23,7 @@
 		creationMenu="<%= publicationsDisplayContext.getCreationMenu() %>"
 		fdsActionDropdownItems="<%= publicationsDisplayContext.getFDSActionDropdownItems() %>"
 		id="<%= PublicationsFDSNames.PUBLICATIONS_ONGOING %>"
-		propsTransformer="{DeletePublicationsFDSPropsTransformer} from change-tracking-web"
+		propsTransformer="{OngoingPublicationsFDSPropsTransformer} from change-tracking-web"
 	/>
 </div>
